requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
seaborn==0.11.2
keras==2.8.0
tensorflow==2.8.0  # Keras is now integrated into TensorFlow 2.x
streamlit==1.8.0
tensorflow==2.8.0
keras==2.8.0
opencv-python-headless==4.5.5.62
numpy==1.22.3
Pillow==8.4.0
